# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bestellenbij logistics platform API
servers:
  - url: /api
    description: Base API path
...
security:
  - bearerAuth: []
paths:
...
  /riders:
    get:
      operationId: listRiders
      tags: [riders]
      summary: List riders with current workload counts
      responses:
        "200":
          description: Riders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiderWithWorkload"
    post:
      operationId: createRider
      tags: [riders]
      summary: Create a new rider account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRiderRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiderWithWorkload"
  /riders/{id}:
    patch:
      operationId: updateRider
      tags: [riders]
      summary: Update rider account or availability
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRiderRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiderWithWorkload"
  /riders/me/availability:
    post:
      operationId: setOwnAvailability
      tags: [riders]
      summary: Rider sets their own availability status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetAvailabilityRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiderWithWorkload"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AccountStatus:
      type: string
      enum: [active, suspended]
...
    RiderAvailability:
      type: string
      enum: [offline, online, backup]
...
    RiderWithWorkload:
      type: object
      required:
        - id
        - userId
        - nameCode
        - name
        - email
        - availabilityStatus
        - accountStatus
        - activeOrderCount
        - queuedOrderCount
      properties:
        id:
          type: string
        userId:
          type: string
        nameCode:
          type: string
          minLength: 1
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        availabilityStatus:
          $ref: "#/components/schemas/RiderAvailability"
        accountStatus:
          $ref: "#/components/schemas/AccountStatus"
        activeOrderCount:
          type: integer
        queuedOrderCount:
          type: integer
    CreateRiderRequest:
      type: object
      required: [email, name, nameCode, password]
      properties:
        email:
          type: string
        name:
          type: string
        nameCode:
          type: string
          minLength: 1
        password:
          type: string
        phone:
          type: ["string", "null"]
        availabilityStatus:
          $ref: "#/components/schemas/RiderAvailability"
    UpdateRiderRequest:
      type: object
      properties:
        name:
          type: string
        nameCode:
          type: string
          minLength: 1
        phone:
          type: ["string", "null"]
        availabilityStatus:
          $ref: "#/components/schemas/RiderAvailability"
        accountStatus:
          $ref: "#/components/schemas/AccountStatus"
    SetAvailabilityRequest:
      type: object
      required: [availabilityStatus]
      properties:
        availabilityStatus:
          $ref: "#/components/schemas/RiderAvailability"
...